requirements.txt
altgraph==0.10.2
aniso8601==1.1.0
appdirs==1.4.0
bdist-mpkg==0.5.0
bottle==0.12.9
Cerberus==0.9.2
click==6.6
Eve==0.6.4
Events==0.2.1
Flask==0.10.1
Flask-MySQL==1.3
Flask-MySQLdb==0.2.0
Flask-PyMongo==0.4.1
Flask-RESTful==0.3.5
futures==3.0.5
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
jsonpickle==0.9.3
macholib==1.5.1
MarkupSafe==0.23
modulegraph==0.10.4
mongoengine==0.10.6
MonthDelta==0.9.1
mysql==0.0.1
MySQL-python==1.2.5
mysqlclient==1.3.7
numpy==1.8.0
packaging==16.8
peewee==2.8.2
plivo==0.11.3
py2app==0.7.3
pymongo==3.3.0
pyOpenSSL==0.13.1
pyparsing==2.1.10
python-dateutil==1.5
pytz==2013.7
requests==2.12.4
scipy==0.13.0
simplejson==3.8.2
six==1.10.0
suds==0.4
Werkzeug==0.11.3
xattr==0.6.4
xmltodict==0.10.2
zope.interface==4.1.1
